BITMAG-992 Change MediatorCleanupInterval to an XML duration
