+++ bitrepository-integration/src/main/resources/quickstart/conf/audittrailservice/ReferenceSettings.xml
@@ -27,7 +27,7 @@
   </GeneralSettings>
 
   <ClientSettings>
-    <MediatorCleanupInterval>1000</MediatorCleanupInterval>
+    <MediatorCleanupInterval>PT1S</MediatorCleanupInterval>
     <ConversationTimeout>3600000</ConversationTimeout>
   </ClientSettings>
